Move mcp_server into src package
